requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-aws
boto3
pydantic
python-dotenv
reportlab
pymongo
python-dotenv
passlib[bcrypt]
langchain
langchain-aws
langchain-google-genai
python-dotenv
